PD-233 학원 상세 페이지를 수정 페이지로 변경
