+++ src/page/academy/[academyId]/index.tsx
@@ -1,125 +1,336 @@
+import { zodResolver } from "@hookform/resolvers/zod";
 import {
-    Container,
-    Typography,
-    Paper,
-    Box,
-    Button,
-    Divider,
-    CircularProgress,
-  } from "@mui/material";
-  import { useQuery } from "@tanstack/react-query";
-  import { isNil } from "lodash-es";
-  import { useEffect } from "react";
-  import { useNavigate, useParams } from "react-router-dom";
-  
-  import { academyApi } from "@/api/academy/request";
-  import { AcademyDetail } from "@/type/academy";
-  
-  const AcademyDetailPage = () => {
-    const navigate = useNavigate();
-    const { academyId } = useParams();
-  
-    const { data: academy, isLoading, error } = useQuery<AcademyDetail>({
-      queryKey: ["academyDetail", academyId],
-      queryFn: () => academyApi.getDetail(Number(academyId)),
-      enabled: !isNil(academyId),
-    });
-  
-    useEffect(() => {
-      if (error) {
-        navigate("/academy");
+  Container,
+  Typography,
+  TextField,
+  Button,
+  Box,
+  CircularProgress,
+  MenuItem,
+  Select,
+  InputLabel,
+  FormControl,
+  Checkbox,
+  FormControlLabel,
+} from "@mui/material";
+import { useQuery, useMutation } from "@tanstack/react-query";
+import { useEffect, useState } from "react";
+import { useForm } from "react-hook-form";
+import { useNavigate, useParams } from "react-router-dom";
+import { toast } from "react-toastify";
+
+import { academyApi } from "@/api/academy/request";
+import {CreateAcademySchema,type CreateAcademyInput,} from "@/type/academy";
+import { LocationTypes } from "@/type/code";
+
+const EditAcademy = () => {
+  const navigate = useNavigate();
+  const { academyId } = useParams();
+  const [selectedFiles, setSelectedFiles] = useState<File[]>([]);
+  const [previewImages, setPreviewImages] = useState<string[]>([]);
+
+  const { data: academy, isLoading, error } = useQuery({
+    queryKey: ["academyDetail", academyId],
+    queryFn: () => academyApi.getDetail(Number(academyId)),
+    enabled: !!academyId,
+  });
+
+  const {
+    register,
+    handleSubmit,
+    formState: { errors },
+    watch,
+    setValue,
+  } = useForm<CreateAcademyInput>({
+    resolver: zodResolver(CreateAcademySchema),
+    defaultValues: {
+      name: "",
+      nameEn: "",
+      representativeName: "",
+      representativeEmail: "",
+      description: "",
+      locationType: "SEOUL",
+      detailedAddress: "",
+      lat: 0,
+      lng: 0,
+      forKindergarten: false,
+      forElementary: false,
+      forMiddleSchool: false,
+      forHighSchool: false,
+      forAdult: false,
+    }
+  });
+
+  // academy 데이터가 로드되면 폼 값 설정
+  useEffect(() => {
+    if (academy) {
+      setValue("name", academy.name);
+      setValue("nameEn", academy.nameEn);
+      setValue("representativeName", academy.representativeName);
+      setValue("representativeEmail", academy.representativeEmail);
+      setValue("description", academy.description || "");
+      setValue("locationType", academy.locationType);
+      setValue("detailedAddress", academy.detailedAddress);
+      setValue("lat", academy.lat ?? 0);
+      setValue("lng", academy.lng ?? 0);
+      setValue("forKindergarten", academy.forKindergarten ?? false);
+      setValue("forElementary", academy.forElementary ?? false);
+      setValue("forMiddleSchool", academy.forMiddleSchool ?? false);
+      setValue("forHighSchool", academy.forHighSchool ?? false);
+      setValue("forAdult", academy.forAdult ?? false);
+      
+      // 기존 이미지 URL을 미리보기에 추가
+      if (academy.imageUrls) {
+        setPreviewImages(academy.imageUrls);
       }
-    }, [academyId, error, navigate]);
-  
-    if (isLoading) {
-      return (
-        <Box display="flex" justifyContent="center" mt={4}>
-          <CircularProgress />
-        </Box>
-      );
     }
-  
-    if (!academy) {
-      return (
-        <Box display="flex" justifyContent="center" mt={4}>
-          <Typography>학원을 찾을 수 없습니다.</Typography>
-        </Box>
-      );
+  }, [academy, setValue]);
+
+  const updateAcademy = useMutation({
+    mutationFn: (data: CreateAcademyInput) => academyApi.updateAcademy(Number(academyId), data),
+    onSuccess: () => {
+      toast.success("학원이 성공적으로 수정되었습니다.");
+      navigate("/academy");
+    },
+    onError: () => {
+      toast.error("학원 수정에 실패했습니다.");
+    },
+  });
+
+  const onSubmit = (data: CreateAcademyInput) => {
+    if (selectedFiles.length > 0) {
+      data.images = selectedFiles;
     }
-  
+    updateAcademy.mutate(data);
+  };
+
+  const handleImageChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    if (event.target.files) {
+      const fileList = Array.from(event.target.files);
+      setSelectedFiles(fileList);
+
+      // 미리보기 이미지 URL 생성 (기존 이미지는 제거하고 새 이미지만 표시)
+      const previewUrls = fileList.map(file => URL.createObjectURL(file));
+      setPreviewImages(previewUrls);
+    }
+  };
+
+  if (isLoading) {
     return (
-      <Container maxWidth="lg" sx={{ mt: 4 }}>
-        <Paper sx={{ p: 3 }}>
-          <Typography variant="h4" component="h1" gutterBottom>
-            {academy.name} ({academy.nameEn})
-          </Typography>
-          <Box sx={{ mb: 2, display: "flex", justifyContent: "space-between" }}>
-            <Typography variant="body2" color="text.secondary">
-              대표자: {academy.representativeName}
-            </Typography>
-            <Typography variant="body2" color="text.secondary">
-              이메일: {academy.representativeEmail}
-            </Typography>
-          </Box>
-          <Divider sx={{ my: 2 }} />
-          <Typography variant="body1" sx={{ whiteSpace: "pre-line", mb: 3 }}>
-            {academy.description || "설명이 없습니다."}
-          </Typography>
-          <Typography variant="body2" sx={{ mb: 1 }}>
-            사업자 등록번호: {academy.businessRegistrationNumber}
-          </Typography>
-          <Typography variant="body2" sx={{ mb: 1 }}>
-            지역: {academy.locationType} / {academy.detailedAddress}
-          </Typography>
-          <Typography variant="body2" sx={{ mb: 1 }}>
-            위도: {academy.lat}, 경도: {academy.lng}
-          </Typography>
-          <Typography variant="body2" sx={{ mb: 1 }}>
-            대상:{" "}
-            {[
-              academy.forKindergarten && "유아",
-              academy.forElementary && "초등학생",
-              academy.forMiddleSchool && "중학생",
-              academy.forHighSchool && "고등학생",
-              academy.forAdult && "성인",
-            ]
-              .filter(Boolean)
-              .join(", ")}
-          </Typography>
-          <Typography variant="body2" sx={{ mb: 1 }}>
-            이미지: {academy.imageUrls?.length || 0}개
-           </Typography>
-           {academy.imageUrls && academy.imageUrls.length > 0 && (
-           <Box
-               sx={{
-               display: "flex",
-               flexWrap: "wrap",
-               gap: 2,
-               mb: 2,
-               }}
-           >
-               {academy.imageUrls.map((url, index) => (
-               <Box key={index} sx={{ width: 150, border: "1px solid #ccc", overflow: "hidden" }}>
-                   <img
-                   src={"https://d1zl1w0yhwh5x4.cloudfront.net/" + url}
-                   alt={`학원 이미지 ${index + 1}`}
-                   style={{ width: "100%", height: "100%", objectFit: "cover" }}
-                   />
-               </Box>
-               ))}
-           </Box>
-           )}
-          <Box sx={{ display: "flex", gap: 2, justifyContent: "flex-end", mt: 3 }}>
-            <Button variant="outlined" onClick={() => navigate("/academy")}>
-              목록
-            </Button>
-            <Button variant="contained" onClick={() => navigate(`/academy/${academyId}/edit`)}>
-              수정
-            </Button>
-          </Box>
-        </Paper>
+      <Container sx={{ display: "flex", justifyContent: "center", mt: 4 }}>
+        <CircularProgress />
       </Container>
     );
-  };
-  
-  export default AcademyDetailPage;
+  }
+  if (error) {
+    return (
+      <Container sx={{ mt: 4 }}>
+        <Typography color="error">학원 정보를 불러오는데 실패했습니다.</Typography>
+      </Container>
+    );
+  }
+
+  return (
+    <Container maxWidth="md">
+      <Typography variant="h4" component="h1" gutterBottom>
+        학원 수정
+      </Typography>
+
+      <Box
+        component="form"
+        onSubmit={handleSubmit(onSubmit)}
+        sx={{ display: "flex", flexDirection: "column", gap: 2 }}
+      >
+        <TextField
+          label="학원명 (국문)"
+          {...register("name")}
+          error={!!errors.name}
+          helperText={errors.name?.message}
+        />
+        <TextField
+          label="학원명 (영문)"
+          {...register("nameEn")}
+          error={!!errors.nameEn}
+          helperText={errors.nameEn?.message}
+        />
+        <TextField
+          label="대표자명"
+          {...register("representativeName")}
+          error={!!errors.representativeName}
+          helperText={errors.representativeName?.message}
+        />
+        <TextField
+          label="대표자 이메일"
+          {...register("representativeEmail")}
+          error={!!errors.representativeEmail}
+          helperText={errors.representativeEmail?.message}
+        />
+        <TextField
+          label="설명"
+          multiline
+          rows={4}
+          {...register("description")}
+          error={!!errors.description}
+          helperText={errors.description?.message}
+        />
+        <FormControl fullWidth error={!!errors.locationType}>
+          <InputLabel id="locationType-label">지역</InputLabel>
+          <Select
+            labelId="locationType-label"
+            label="지역"
+            value={watch("locationType") || ""}
+            {...register("locationType")}
+          >
+            {LocationTypes.map((type) => (
+              <MenuItem key={type} value={type}>
+                {type}
+              </MenuItem>
+            ))}
+          </Select>
+        </FormControl>
+
+        <TextField
+          label="상세 주소"
+          {...register("detailedAddress")}
+          error={!!errors.detailedAddress}
+          helperText={errors.detailedAddress?.message}
+        />
+        <TextField
+          label="위도"
+          type="number"
+          inputProps={{ step: "any" }}
+          {...register("lat", {
+            valueAsNumber: true,
+            validate: (value) =>
+              !isNaN(value) || "유효한 숫자를 입력해주세요",
+          })}
+          error={!!errors.lat}
+          helperText={errors.lat?.message}
+        />
+        <TextField
+          label="경도"
+          type="number"
+          inputProps={{ step: "any" }}
+          {...register("lng", {
+            valueAsNumber: true,
+            validate: (value) =>
+              !isNaN(value) || "유효한 숫자를 입력해주세요",
+          })}
+          error={!!errors.lng}
+          helperText={errors.lng?.message}
+        />
+
+        <Box display="flex" flexDirection="column" gap={1}>
+          <FormControlLabel
+            control={<Checkbox 
+              {...register("forKindergarten")} 
+              checked={watch("forKindergarten")} 
+            />}
+            label="유아 대상"
+          />
+          <FormControlLabel
+            control={<Checkbox 
+              {...register("forElementary")} 
+              checked={watch("forElementary")}
+            />}
+            label="초등학생 대상"
+          />
+          <FormControlLabel
+            control={<Checkbox 
+              {...register("forMiddleSchool")} 
+              checked={watch("forMiddleSchool")}
+            />}
+            label="중학생 대상"
+          />
+          <FormControlLabel
+            control={<Checkbox 
+              {...register("forHighSchool")} 
+              checked={watch("forHighSchool")} 
+            />}
+            label="고등학생 대상"
+          />
+          <FormControlLabel
+            control={<Checkbox 
+              {...register("forAdult")} 
+              checked={watch("forAdult")}
+            />}
+            label="성인 대상"
+          />
+        </Box>
+
+        <Button variant="outlined" component="label">
+          이미지 업로드
+          <input
+            type="file"
+            hidden
+            multiple
+            accept="image/*"
+            onChange={handleImageChange}
+          />
+        </Button>
+        <Typography variant="body2">
+          선택한 새 파일 수: {selectedFiles.length}
+        </Typography>
+
+        {/* 기존 이미지 */}
+        {academy?.imageUrls && academy.imageUrls.length > 0 && (
+          <Box>
+            <Typography variant="subtitle1" gutterBottom>
+              기존 이미지
+            </Typography>
+            <Box sx={{ display: 'flex', gap: 2, flexWrap: 'wrap' }}>
+              {academy.imageUrls.map((url, index) => (
+                <Box key={index} sx={{ width: 150, border: "1px solid #ccc", overflow: "hidden" }}>
+                  <img
+                    src={`https://d1zl1w0yhwh5x4.cloudfront.net/${url}`}
+                    alt={`기존 이미지 ${index + 1}`}
+                    style={{ width: "100%", height: "100%", objectFit: "cover" }}
+                  />
+                </Box>
+              ))}
+            </Box>
+          </Box>
+        )}
+
+        {/* 새로 선택한 이미지 미리보기 */}
+        {previewImages.length > 0 && (
+          <Box>
+            <Typography variant="subtitle1" gutterBottom>
+              새로 선택한 이미지
+            </Typography>
+            <Box sx={{ display: 'flex', gap: 2, flexWrap: 'wrap' }}>
+              {previewImages.map((previewUrl, index) => (
+                <Box key={index} sx={{ width: 150, border: "1px solid #ccc", overflow: "hidden" }}>
+                  <img
+                    src={previewUrl}
+                    alt={`새 이미지 ${index + 1}`}
+                    style={{ width: "100%", height: "100%", objectFit: "cover" }}
+                  />
+                </Box>
+              ))}
+            </Box>
+          </Box>
+        )}
+
+        <Box display="flex" gap={2} mt={2}>
+          <Button variant="outlined" onClick={() => navigate("/academy")}>
+            취소
+          </Button>
+          <Button
+            type="submit"
+            variant="contained"
+            disabled={updateAcademy.isPending}
+          >
+            {updateAcademy.isPending ? (
+              <CircularProgress size={24} />
+            ) : (
+              "수정"
+            )}
+          </Button>
+        </Box>
+      </Box>
+    </Container>
+  );
+};
+
+export default EditAcademy;
